Migrate user IDs to BIGINT; add profile images
Change integer-based user/post IDs to BIGINT across DB schema and initialization scripts (movie_reviews, post_views, password_resets, watchlist, tv_series_reviews, chat/notifications, etc.) and add ALTER TABLE migrations to convert existing columns. Add new user profile fields (profile_photo_url, banner_url, profile_photo_position, banner_position) in initDatabase and flix_db.sql and update existing rows; update updateMyProfile controller to accept and persist these fields. Frontend: introduce buildApiUrl helper and replace direct VITE_API_URL usage, add ImageCropperModal component and styles (react-easy-crop dependency), wire a new profile-image upload route and include uploaded sample images. Misc: watchlist and MoviesPage UI/logic improvements for saving state, and assorted CSS updates for profile editing UI.

+++ backend/sql/movie_reviews.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS flix.movie_reviews (
   id_review SERIAL PRIMARY KEY,
   tmdb_movie_id INTEGER NOT NULL,
-  id_user INTEGER NOT NULL REFERENCES flix.users(id_user) ON DELETE CASCADE,
+  id_user BIGINT NOT NULL REFERENCES flix.users(id_user) ON DELETE CASCADE,
   parent_review_id INTEGER REFERENCES flix.movie_reviews(id_review) ON DELETE CASCADE,
   content TEXT NOT NULL,
   rating SMALLINT,
@@ -20,7 +20,7 @@ ON flix.movie_reviews (parent_review_id);
 CREATE TABLE IF NOT EXISTS flix.movie_review_likes (
   id_like SERIAL PRIMARY KEY,
   id_review INTEGER NOT NULL REFERENCES flix.movie_reviews(id_review) ON DELETE CASCADE,
-  id_user INTEGER NOT NULL REFERENCES flix.users(id_user) ON DELETE CASCADE,
+  id_user BIGINT NOT NULL REFERENCES flix.users(id_user) ON DELETE CASCADE,
   created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
   CONSTRAINT uq_movie_review_likes UNIQUE (id_review, id_user)
 );
