+++ backend/sql/post_views.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS flix.post_views (
   id_view SERIAL PRIMARY KEY,
-  id_post INTEGER NOT NULL REFERENCES flix.posts(id_post) ON DELETE CASCADE,
-  id_user INTEGER NOT NULL REFERENCES flix.users(id_user) ON DELETE CASCADE,
+  id_post BIGINT NOT NULL REFERENCES flix.posts(id_post) ON DELETE CASCADE,
+  id_user BIGINT NOT NULL REFERENCES flix.users(id_user) ON DELETE CASCADE,
   viewed_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
   CONSTRAINT post_views_unique_user_post UNIQUE (id_post, id_user)
 );
